Add instructions for local setup on MacOS
Recently, OpenPrescribing was set up on my M2 MacBook. We thought it would be worth documenting the steps taken, as some parts of the README are outdated or linux-specific.

I have taken the approach of creating DEVELOPERS.md and putting a MacOS section in it, and have tried to make it such that someone can use this doc only to do the local setup and not have to cross-reference with the readme. Hence, a handful of sentences in the README are repeated here.