Add a function that reads a CSV file of concessions
- Occasionally, price concessions might be announced
on a different part of the Community Pharmacy England
than the part that we scrape.

- We would like to incorporate these concessions by
manually maintaining a CSV file. Therefore, write a function
that can parse such a file and produce a list of dictionaries
that are in the same format as a list of scraped concessions.

- Raise an AssertionError if a provided CSV file of
manually added price concessions contains
a date that is not the first of a month,
and add a test for this.