Look up discounts for all presentations at once
This makes quite a significant difference vs doing separate queries to
get the discounts for each substitution set; possibly up to a 50%
speed-up.

This only affects requests with a cold cache, but we've been having
problems where requests timeout before they can write to the cache so
anything which might get us under the timeout is potentially a big win.