Bump gunicorn's limit-request-line setting
This setting controls the maximum size of HTTP request lines.  For some
measures (eg silver dressings), the link to the analyse page is long
enough to go over the default.  8190 is the maximum we can set it
without setting it to be unlimited.

More: https://docs.gunicorn.org/en/stable/settings.html#limit-request-line